% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adjmat_to_equal}
\alias{adjmat_to_equal}
\title{adjmat_to_equal}
\arguments{
\item{adjmat}{Matrix identifying which nodes are neighbors (are linked).}

\item{maxiter}{Integer giving the maximum number of attempts to connect two nodes.}
}
\description{
Take a (graph theory) adjacency matrix and build equivalence classes.
Two nodes are in the same class if there is a path between them.
Returns vector. Each position represents a node and nodes with the same value are part of the same equivalence class.
Implicitally assumes that edges are bidirectional (a path in one direction leads to a path in the other)
}
\keyword{internal}
