% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_switching.r
\name{apply_swaps_helper}
\alias{apply_swaps_helper}
\title{Applies label swapping based on an output from identify_swaps_helper.}
\usage{
apply_swaps_helper(dlcm, identify_out, nwarmup)
}
\arguments{
\item{dlcm}{An dependent latent class model from dependentLCM_fit().}

\item{identify_out}{Output from identify_swaps_helper.}

\item{nwarmup}{integer. Which iterations are warmup iterations?}
}
\description{
Applies label swapping based on an output from identify_swaps_helper.
}
\keyword{internal}
