% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Archive::domains2mat}
\alias{Archive::domains2mat}
\title{Archive::domains2mat
Helper function for Archive::add
Used to convert the domains from map form to matrix form for storage
Although we could make a deep copy of the domain map each time, this would be unproductive because we need it in matrix form later for R. Therfore we convert to matrix.}
\description{
Archive::domains2mat
Helper function for Archive::add
Used to convert the domains from map form to matrix form for storage
Although we could make a deep copy of the domain map each time, this would be unproductive because we need it in matrix form later for R. Therfore we convert to matrix.
}
\keyword{internal}
