% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Archive::set_initial}
\alias{Archive::set_initial}
\title{Archive::set_initial
Initializes the archive (namely reserving memory)}
\description{
Archive::set_initial
Initializes the archive (namely reserving memory)
}
\keyword{internal}
