% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesContainer::run}
\alias{BayesContainer::run}
\title{BayesContainer::run
Does #nitr MCMC steps on all bayes parameters}
\description{
BayesContainer::run
Does #nitr MCMC steps on all bayes parameters
}
\keyword{internal}
