% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesParameter::class_lprob}
\alias{BayesParameter::class_lprob}
\title{BayesParameter::class_lprob}
\arguments{
\item{xclass}{The class this observation is (assumed to be) in}

\item{xobs}{The response pattern we are investigation.}
}
\description{
What is the (log) probability that this response pattern was generated by this class?

What is the (log) probability that this response pattern was generated by EACH class?
Assumes all parameters are fixed/known.
}
\keyword{internal}
