% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesParameter::class_pi_args}
\alias{BayesParameter::class_pi_args}
\title{BayesParameter::class_pi_args
Calculate the dirichlet parameters for the posterior of pi (pi used for classes)}
\description{
BayesParameter::class_pi_args
Calculate the dirichlet parameters for the posterior of pi (pi used for classes)
}
\keyword{internal}
