% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesParameter::classes_next}
\alias{BayesParameter::classes_next}
\title{BayesParameter::classes_next
Do gibbs sampling to calculate the class of each observation}
\description{
BayesParameter::classes_next
Do gibbs sampling to calculate the class of each observation
}
\keyword{internal}
