% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesParameter::domain_accept}
\alias{BayesParameter::domain_accept}
\title{BayesParameter::domain_accept}
\description{
In metroplis algorithm when choosing to update domains.
We examine the proposal and decide whether to accept it.
}
\keyword{internal}
