% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesParameter::domain_addCount}
\alias{BayesParameter::domain_addCount}
\title{BayesParameter::domain_addCount
Add given observation to all the counts in all provided domains}
\arguments{
\item{domains}{map of domains we wish to add this pattern to}

\item{xobs}{the response pattern we are counting}

\item{xlcass}{The class of this observation. Needed because different domains correspond to different classes.}
}
\description{
BayesParameter::domain_addCount
Add given observation to all the counts in all provided domains

BayesParameter::domain_addCount
Add given observation to all the counts in all of this BayesParameter's domains
}
\keyword{internal}
