% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesParameter::domain_addCounts}
\alias{BayesParameter::domain_addCounts}
\title{BayesParameter::domain_addCounts}
\arguments{
\item{domains}{Map of domains holding counts}

\item{x}{Matrix of the response patterns}

\item{reset_counts}{True if we should set all counts to zero before counting x.}
}
\description{
Add given observations to counts in all of this given domains

Assumes the classes of x correspond to the classes in BayesParameter.

Add given observations to counts in all domains in BayesParameter
}
\keyword{internal}
