% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesParameter::domain_getloglik_x}
\alias{BayesParameter::domain_getloglik_x}
\title{BayesParameter::domain_getloglik_x
For a given domain, we want to know the probability of observing a series of responses.
We calculate this probability conditioning on parameters, but marginalizing over theta (and convert to log scale)}
\arguments{
\item{pattern_counts}{The counts for each pattern in this domain (vector).}

\item{theta_alpha}{Hyperparameter describing the Dirichlet concentration parameters for the theta prior.}
}
\description{
BayesParameter::domain_getloglik_x
For a given domain, we want to know the probability of observing a series of responses.
We calculate this probability conditioning on parameters, but marginalizing over theta (and convert to log scale)
}
\keyword{internal}
