% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesParameter::domain_next}
\alias{BayesParameter::domain_next}
\title{BayesParameter::domain_next
We use a metropolis algorithm to try to update domains
In essence we choose one item at random and either move it to another domain or swap it with an item from another domain}
\description{
BayesParameter::domain_next
We use a metropolis algorithm to try to update domains
In essence we choose one item at random and either move it to another domain or swap it with an item from another domain
}
\keyword{internal}
