% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesParameter::domain_proposal}
\alias{BayesParameter::domain_proposal}
\title{BayesParameter::domain_proposal
Propose new domains in metropolis algorithm
Essentially we choose two domains at random and move items between them (oversimplification).
We take a nonempty domain. 1) With some probability we move one of its items to an empty domain. Otherwise we choose a second nonempty domain. 2) With some probability we swap a pair of items between the two nonempty domains. 3) Otherwise we move one item from the first domain into the second.
There are some restrictions including 1) We do not swap items if both domains are singletons (this would cause no change up to relabeling), and enforce maximum number of items per domain given in hparams.}
\description{
BayesParameter::domain_proposal
Propose new domains in metropolis algorithm
Essentially we choose two domains at random and move items between them (oversimplification).
We take a nonempty domain. 1) With some probability we move one of its items to an empty domain. Otherwise we choose a second nonempty domain. 2) With some probability we swap a pair of items between the two nonempty domains. 3) Otherwise we move one item from the first domain into the second.
There are some restrictions including 1) We do not swap items if both domains are singletons (this would cause no change up to relabeling), and enforce maximum number of items per domain given in hparams.
}
\keyword{internal}
