% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesParameter::domains_next}
\alias{BayesParameter::domains_next}
\title{BayesParameter::domains_next
Use metropolis algorithm to update domains. Repeat # of times set by hparams.}
\description{
BayesParameter::domains_next
Use metropolis algorithm to update domains. Repeat # of times set by hparams.
}
\keyword{internal}
