% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesParameter::set_class_loglik}
\alias{BayesParameter::set_class_loglik}
\title{BayesParameter::set_class_loglik}
\arguments{
\item{x}{Matrix of responses}

\item{reset}{Whether to initialize the matrix dimensions before running}
}
\description{
For each observation what is the (log) probability of seeing its response pattern under EACH class?
Updates corresponding property in BayesParameter instance.
Assumes all parameters are fixed/known.
}
\keyword{internal}
