% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesParameter::set_class_loglik_collapsed}
\alias{BayesParameter::set_class_loglik_collapsed}
\title{BayesParameter::set_class_loglik_collapsed}
\arguments{
\item{x}{Matrix of responses}

\item{hparams}{Hyperparameter values}
}
\description{
For each observation what is the collapsed (log) probability of this observation belonging to this class? Collapsed on theta and pi.
Updates corresponding property in BayesParameter instance. Only correct up to proportionality.
Assumes all parameters are fixed/known.
}
\keyword{internal}
