% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesParameter::thetas_next}
\alias{BayesParameter::thetas_next}
\title{BayesParameter::thetas_next
Do gibbs sampling to calculate thetas for each domain}
\description{
BayesParameter::thetas_next
Do gibbs sampling to calculate thetas for each domain
}
\keyword{internal}
