% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{check_params}
\alias{check_params}
\title{Check parameters for internal consistency}
\usage{
check_params(all_params)
}
\arguments{
\item{all_params}{list. List of parameters and hyperparameters}
}
\description{
Check parameters for internal consistency
}
\keyword{internal}
