% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDLCM.r
\name{dependence_intensity}
\alias{dependence_intensity}
\title{For each domain domain, compare the probabilities under local dependence (DLCM) versus under local independence (traditional LCM). Describes the amount of dependence captured by our DLCM model.}
\usage{
dependence_intensity(thetas_avg, dlcm, items_ids = NULL)
}
\arguments{
\item{thetas_avg}{dataframe. Describes average response probabilities for different domains. As returned from dlcm.summary()}

\item{dlcm}{list. Fitted dependent LCM}

\item{items_ids}{integerVector Which domains do we want to process?}
}
\value{
DataFrames with:
\itemize{
\item{"prob_dependent"}{= Probability under local dependence. Same as input 'probs'.}
\item{"prob_marginal"}{= Probability of this response pattern under local independence.}
\item{"odds_ratio"}{= Compares prob_dependent and prob_marginal. Values far from '1' show strong dependence.}
\item{"odds_ratio_str"}{= odds_ratio as a fraction}
}
}
\description{
For each domain domain, compare the probabilities under local dependence (DLCM) versus under local independence (traditional LCM). Describes the amount of dependence captured by our DLCM model.
}
