% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDLCM.r
\name{dependence_intensity_marginals}
\alias{dependence_intensity_marginals}
\title{Takes a single domain. Calculates teh marginal probabilities}
\usage{
dependence_intensity_marginals(values, probs)
}
\arguments{
\item{values}{characterMatrix One row for each response pattern to this domain. One column per item. Cell contains the value of a given item under the given response pattern.}

\item{probs}{numericVector. For a single class, probability of returning this response pattern.}
}
\value{
Marginal probabilities as part of a dataframe
}
\description{
Takes a single domain. Calculates teh marginal probabilities
}
\keyword{internal}
