% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDLCM.r
\name{dependence_intensity_one}
\alias{dependence_intensity_one}
\title{Takes a single domain. Compare the probabilites under the given probabilities versus under local independence.}
\usage{
dependence_intensity_one(values, probs)
}
\arguments{
\item{values}{characterMatrix One row for each response pattern to this domain. One column per item. Cell contains the value of a given item under the given response pattern.}

\item{probs}{numericVector. For a single class, probability of returning this response pattern.}
}
\value{
DataFrames with:
\itemize{
\item{"prob_dependent"}{= Probability under local dependence. Same as input 'probs'.}
\item{"prob_marginal"}{= Probability of this response pattern under local independence.}
\item{"odds_ratio"}{= Compares prob_dependent and prob_marginal. Values far from '1' show strong dependence.}
\item{"odds_ratio_str"}{= odds_ratio as a fraction}
}
}
\description{
Takes a single domain. Compare the probabilites under the given probabilities versus under local independence.
}
\keyword{internal}
