% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDLCM.r
\name{dependence_intensity_ratios}
\alias{dependence_intensity_ratios}
\title{Takes a single domain. Compares dependent versus marginal probabilities of each pattern.}
\usage{
dependence_intensity_ratios(probs, prob_marginal)
}
\arguments{
\item{probs}{numericVector. For a single class, probability of returning this response pattern.}

\item{probs_marginal}{numericVector. For a single class under conditonal independence, probability of returning this response pattern.}
}
\description{
Takes a single domain. Compares dependent versus marginal probabilities of each pattern.
}
\keyword{internal}
