% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\docType{package}
\name{dependentLCM}
\alias{dependentLCM}
\title{dependentLCM: Dependent Latent Class Model}
\description{
Latent Class Models (LCMs) are used to cluster multivariate categorical data (e.g. group participants based on survey responses). Traditional LCMs assume a property called conditional independence. This assumption can be restrictive, leading to model misspecification and overparameterization. To combat this problem, we developed a novel Bayesian model called a Dependent Latent Class Model (DLCM), which permits conditional dependence. Compared to traditional LCMs, DLCMs are effective in applications with time series, overlapping items, and structural zeroes.

The primary function is dependentLCM_fit. 

Bowers, J., & Culpepper, S. (2022). Dependent Latent Class Models (Version 1). arXiv. \url{https://doi.org/10.48550/ARXIV.2205.08677}

\url{https://arxiv.org/abs/2205.08677}

\url{https://github.com/jessebowers/dependentLCM}
}
