% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{dependentLCM_fit}
\alias{dependentLCM_fit}
\title{Fits a bayesian dependent LCM model}
\usage{
dependentLCM_fit(
  nitr,
  df = NULL,
  mat = NULL,
  nclass = NCLASS,
  ndomains = NULL,
  class2domain = NULL,
  classPi_alpha = CLASSPI_ALPHA,
  domain_alpha = NULL,
  domain_maxitems = NULL,
  theta_alpha = THETA_ALPHA,
  domain_proposal_empty = DOMAIN_PROPOSAL_EMPTY,
  domain_proposal_swap = DOMAIN_PROPOSAL_SWAP,
  domain_nproposals = NULL,
  class_pi = NULL,
  classes = NULL,
  thetas = NULL,
  class_init_method = "kmodes"
)
}
\arguments{
\item{nitr}{integer. Number of iterations to run the bayes MCMC}

\item{df}{dataframe. Raw data you are fitting with all assumed to be factors}

\item{mat}{matrix. The data you wish to analyze}

\item{nclass}{integer. Number of subject latent classes}

\item{ndomains}{integer. Number of item domains}

\item{class2domain}{integer vector of length nclass. Classes with same value have same domains.}

\item{classPi_alpha}{numeric vector. Bayes hyperparameter giving prior for Pi.}

\item{domain_alpha}{numeric. Bayes hyperparameter giving prior for domains.}

\item{domain_maxitems}{iinteger. Maximum number of items which can be in a doamin.}

\item{theta_alpha}{numeric. Bayes hyperparemter giving the prior for theta/probabilties.}

\item{domain_proposal_empty}{numeric. Sets how often the domain metropolis proposal function pairs a nonempty domain with an empty domain.}

\item{domain_proposal_swap}{numeric. Sets how often the domain metropolis proposal function swaps items between domains.}

\item{class_pi}{numeric vector size nclass. Initial condition for bayes parameter pi.}

\item{classes}{integer vector size nrow(df). Initial condition for subject classes.}

\item{thetas}{list. Initial values for thetas/probabilites.}

\item{class_init_method}{string. Decides how 'classes' is defaulted if NULL. One of "kmodes" or "random"}
}
\description{
Fits a bayesian dependent LCM model
}
