% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{dlcm.get_waic}
\alias{dlcm.get_waic}
\title{Calculate likelihood and WAIC}
\usage{
dlcm.get_waic(this_sim, itrs = NULL, alpha = 0)
}
\arguments{
\item{itrs}{integer vector. Which iterations should be include in calculation?}

\item{alpha}{percentage. Used for robust calculation of varaince in WAIC. Ignore the alpha most extreme iterations.}

\item{this_sim.}{Dependent Latent class model}
}
\description{
Calculate likelihood and WAIC
}
