% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{dlcm.get_waic_fromagg}
\alias{dlcm.get_waic_fromagg}
\title{Calculate likelihood and WAIC from aggregated information. See dlcm.get_waic() for details.}
\usage{
dlcm.get_waic_fromagg(dlcm)
}
\description{
Calculate likelihood and WAIC from aggregated information. See dlcm.get_waic() for details.
}
\keyword{internal}
