% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDLCM.r
\name{dlcm.get_waic_fromraw}
\alias{dlcm.get_waic_fromraw}
\title{Calculate likelihood and WAIC from raw class_loglik. See dlcm.get_waic() for details.
Warning: Only works if class_loglik is saved, as in: dependentLCM_fit(..., save_itrs=c(class_loglik=Inf, ...))}
\usage{
dlcm.get_waic_fromraw(dlcm, itrs = NULL)
}
\description{
Calculate likelihood and WAIC from raw class_loglik. See dlcm.get_waic() for details.
Warning: Only works if class_loglik is saved, as in: dependentLCM_fit(..., save_itrs=c(class_loglik=Inf, ...))
}
\keyword{internal}
