% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DomainCount::countReset}
\alias{DomainCount::countReset}
\title{DomainCount::countReset}
\description{
Set counts to zero. 
Counts measure the number of times each pattern appears in the data.
}
\keyword{internal}
