% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DomainCount::countSubtract}
\alias{DomainCount::countSubtract}
\title{DomainCount::countSubtract}
\arguments{
\item{xobs}{One FULL response pattern (not just the items in this domain)}
}
\description{
Removes one observation to this domain's counts.
}
\keyword{internal}
