% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DomainCount::drop_item}
\alias{DomainCount::drop_item}
\title{DomainCount::drop_item}
\arguments{
\item{item}{item you wish to remove}
}
\description{
Efficiently remove item from domain (without needing to re-calculate counts)
}
\keyword{internal}
