% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DomainCount::getloglik_marginal}
\alias{DomainCount::getloglik_marginal}
\title{DomainCount::getloglik_marginal}
\arguments{
\item{hparams}{hyperparameters}
}
\description{
For a given domain, we want to know the probability of observing a series of responses.
We calculate this probability conditioning on parameters, but marginalizing over theta (and convert to log scale)
For domain_theta_prior_type="patternadjusted", we also include partial prior for domain
}
\keyword{internal}
