% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DomainCount::getloglik_x}
\alias{DomainCount::getloglik_x}
\title{DomainCount::getloglik_x}
\arguments{
\item{hparams}{hyperparameters}
}
\description{
For a given domain, we want to know the probability of observing a series of responses.
We calculate this probability conditioning on parameters, but marginalizing over theta (and convert to log scale)
}
\keyword{internal}
