% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DomainCount::id2pattern}
\alias{DomainCount::id2pattern}
\title{DomainCount::id2pattern
Convert pattern ID to original response pattern
See utilities ::id2pattern(.) for more details
Differs from ::id2pattern(.) in that pattern2id_map is a cumulative compared to mapvec}
\description{
DomainCount::id2pattern
Convert pattern ID to original response pattern
See utilities ::id2pattern(.) for more details
Differs from ::id2pattern(.) in that pattern2id_map is a cumulative compared to mapvec
}
\keyword{internal}
