% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DomainCount::itemsid_calc}
\alias{DomainCount::itemsid_calc}
\title{DomainCount::itemsid_calc}
\description{
Convert domain.items into an integer id describing the items in this domain
}
\keyword{internal}
