% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DomainCount::pattern2id}
\alias{DomainCount::pattern2id}
\title{DomainCount::pattern2id}
\description{
Convert (vector) response pattern to representative ID
See id2pattern(.) for more details
}
\keyword{internal}
