% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DomainCount::reduce_items}
\alias{DomainCount::reduce_items}
\title{DomainCount::reduce_items}
\arguments{
\item{items_new}{Resulting items in your domain. Assumed(!!) to be subset of current items.}
}
\description{
Efficiently remove items from domain (without needing to re-calculate counts)
}
\keyword{internal}
