% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DomainCount::set_initial}
\alias{DomainCount::set_initial}
\title{DomainCount::set_initial}
\arguments{
\item{items_in}{Which items are in this domain?}

\item{hyparams}{Hyperparmeters}

\item{lthetas_in}{Log probablities of each response pattern of these items (optional)}

\item{list_domain}{List of all arguments for set_initial()}

\item{hparams}{hyperparmeters}
}
\description{
Set all (non-count) values for this domain

As other set_initial() but with list compatibilities for R
}
\keyword{internal}
