% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DomainCount::theta_alpha_fun}
\alias{DomainCount::theta_alpha_fun}
\title{DomainCount::theta_alpha_fun}
\arguments{
\item{hparams}{hyperparameters}
}
\description{
Allows for theta prior to change as domains get merged
}
\keyword{internal}
