% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM_fit.r
\name{doWarmup}
\alias{doWarmup}
\title{Does *EXTRA* MCMC warmup step if applicable}
\usage{
doWarmup(args, warmup_settings = NULL, warmup_dlcm = NULL)
}
\arguments{
\item{args}{list All arguments given to dependentLCM_fit}

\item{warmup_settings}{list. Optionally allows for an extra warmup cycle using different parameters vs the arguments given for the main process. Any parameters in the given list will overwrite the arguments of the same name in the warmup cycle.}

\item{warmup_dlcm}{list. A past DLCM fit (from dependentLCM_fit). The last iteration of the DLCM is used as the Bayes parameters. Takes precidence over warmup_settings.}
}
\description{
Does *EXTRA* MCMC warmup step if applicable
}
\keyword{internal}
