% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{equal_to_adjmat}
\alias{equal_to_adjmat}
\title{equal_to_adjmat}
\arguments{
\item{eqclass_vec}{Vector describing the equivalence classes. 
Each index represents a separate item and indexes with the same value are in the same equivalence class.}
}
\description{
Convert equivalence classes into a (graph theory style) adjacency matrix.
All elements of the same equivalence class are considered neighbors.
}
\keyword{internal}
