% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{expSumLog}
\alias{expSumLog}
\title{Same as log(sum(exp(x))), but adjusted to improve precision.
Assumes x are logged values. Calculates sum(e^x) and then converts back to log scale}
\usage{
expSumLog(x)
}
\arguments{
\item{x}{numeric vector in log scale}
}
\description{
Same as log(sum(exp(x))), but adjusted to improve precision.
Assumes x are logged values. Calculates sum(e^x) and then converts back to log scale
}
\keyword{internal}
