% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{get_class_probs}
\alias{get_class_probs}
\title{For each iteration calculate the probability that a given observation is in each class.}
\usage{
get_class_probs(dlcm)
}
\arguments{
\item{dlcm}{Dependent latent class model outptut from dependentLCM_fit}
}
\description{
For each iteration calculate the probability that a given observation is in each class.
}
