% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDLCM.r
\name{get_itrs_helper}
\alias{get_itrs_helper}
\title{Generate a range of iterations which are both saved and occur after warmup.}
\usage{
get_itrs_helper(nsaved, nwarmup, nitr)
}
\description{
Generate a range of iterations which are both saved and occur after warmup.
}
\keyword{internal}
