% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDLCM.r
\name{get_jointLikelihood}
\alias{get_jointLikelihood}
\title{For each iteration, calculate the prior probabilities of each parameter and conditional probability of our responses
Warning: Only works if class_loglik is saved, as in: dependentLCM_fit(..., save_itrs=c(class_loglik=Inf, ...))}
\usage{
get_jointLikelihood(dlcm)
}
\arguments{
\item{dlcm}{Dependent latent class model outptut from dependentLCM_fit}
}
\description{
For each iteration, calculate the prior probabilities of each parameter and conditional probability of our responses
Warning: Only works if class_loglik is saved, as in: dependentLCM_fit(..., save_itrs=c(class_loglik=Inf, ...))
}
