% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDLCM.r
\name{get_jointLikelihood_obs}
\alias{get_jointLikelihood_obs}
\title{Calculates probability of each response collapsed on classes, then aggregated for each iteration}
\usage{
get_jointLikelihood_obs(dlcm, method)
}
\arguments{
\item{dlcm}{Domain latent class model outptut from dependentLCM_fit}

\item{method}{Defines what variable from dlcm$mcmc is used to calculate resposne probabilities. One of c("itrLogLik", "class_loglik", "").}
}
\description{
Calculates probability of each response collapsed on classes, then aggregated for each iteration
}
\keyword{internal}
