% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDLCM.r
\name{get_jointLikelihood_priors}
\alias{get_jointLikelihood_priors}
\title{For each iteration, calculate the prior probabilities of each parameter}
\usage{
get_jointLikelihood_priors(dlcm)
}
\arguments{
\item{dlcm}{Domain latent class model outptut from dependentLCM_fit}
}
\description{
For each iteration, calculate the prior probabilities of each parameter
}
\keyword{internal}
