% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_pattern2id_map}
\alias{get_pattern2id_map}
\title{get_pattern2id_map}
\usage{
get_pattern2id_map(item_nlevels, items)
}
\arguments{
\item{item_nlevels}{How many possible values each item can take. The response pattern is assumed to}

\item{items_ids}{ID reprsenting which items (positions) should be 'filled'.}
}
\description{
Create 'conversion vector' for converting (vector) response pattern to representative ID
The dot product of the pattern2id_map and the response pattern gives the pattern id. This does the opposite of id2pattern(.).
}
