% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_which_strs}
\alias{get_which_strs}
\title{get_which_strs}
\usage{
get_which_strs(x)
}
\arguments{
\item{x}{matrix. One observation per column and item (boolean) item per row.}
}
\value{
Returns a vector of strings. For each column, give the indices which are TRUE. For instance column [TRUE,FALSE,TRUE,FALSE,FALSE,TRUE] gives ",1,3,6,".
}
\description{
Helper function. For each column of matrix x, identify which cells are true with a string.
Same as R code: apply(x, 2, function(x) paste0(c("", which(x), ""), collapse=",")). Implemented in C++ to improve speed.
}
\keyword{internal}
