% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{getStart_bayes_params}
\alias{getStart_bayes_params}
\title{Generates initial values for bayes parameters}
\usage{
getStart_bayes_params(
  mat,
  hparams,
  class_pi = NULL,
  classes = NULL,
  domains = NULL,
  class_init_method = CLASS_INIT_METHODS[1],
  warmup_dlcm = NULL
)
}
\arguments{
\item{mat}{matrix. The data you wish to analyze}

\item{hparams}{list. List of hyperparameters from getStart_hparams.}

\item{class_pi}{numeric vector size nclass. Initial condition for bayes parameter pi.}

\item{classes}{integer vector size nrow(df). Initial condition for subject classes.}

\item{domains}{list. Initial values for domains/probabilites.}

\item{class_init_method}{string. Decides how 'classes' is defaulted if NULL. One of "kmodes" or "random" or "random_centers", "random_centers_polar"}

\item{warmup_dlcm}{list. A past DLCM fit (from dependentLCM_fit). The last iteration of the DLCM is used as the Bayes parameters.}
}
\description{
Generates initial values for bayes parameters
}
\keyword{internal}
