% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{getStart_class_kmodes}
\alias{getStart_class_kmodes}
\title{Choose class of each observation using kmodes}
\usage{
getStart_class_kmodes(mat, hparams, iter.max = 2, ...)
}
\arguments{
\item{mat}{matrix. Raw data.}

\item{hparams}{list. List of hyperparameters}
}
\description{
Choose class of each observation using kmodes
}
\keyword{internal}
