% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM_fit.r
\name{getStart_class_random}
\alias{getStart_class_random}
\title{Choose class of each observation entirely at random}
\usage{
getStart_class_random(mat, hparams)
}
\arguments{
\item{mat}{matrix. Raw data.}

\item{hparams}{list. List of hyperparameters}
}
\description{
Choose class of each observation entirely at random
}
\keyword{internal}
