% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{getStart_domains}
\alias{getStart_domains}
\title{Choose starting domain values.}
\usage{
getStart_domains(mat, classes, hparams)
}
\arguments{
\item{mat}{matrix. Raw data.}

\item{classes}{integer vector. The class of each observation.}

\item{hparams}{list. List of hyperparameters}
}
\description{
Choose starting domain values.
}
\keyword{internal}
