% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{getStart_hparams}
\alias{getStart_hparams}
\title{Generate hyperparameters
Generate a list of hyperparameters, adding default values when necessary}
\usage{
getStart_hparams(
  df = NULL,
  nitems = NULL,
  nclass = NCLASS,
  ndomains = NULL,
  class2domain = NULL,
  classPi_alpha = CLASSPI_ALPHA,
  domain_alpha = NULL,
  domain_maxitems = NULL,
  theta_alpha = THETA_ALPHA,
  domain_proposal_empty = DOMAIN_PROPOSAL_EMPTY,
  domain_proposal_swap = DOMAIN_PROPOSAL_SWAP,
  domain_nproposals = NULL
)
}
\arguments{
\item{df}{dataframe. Raw data you are fitting with all assumed to be factors}

\item{nitems}{integer. Number of columns of df.}

\item{nclass}{integer. Number of subject latent classes}

\item{ndomains}{integer. Number of item domains}

\item{class2domain}{integer vector of length nclass. Classes with same value have same domains.}

\item{classPi_alpha}{numeric vector. Bayes hyperparameter giving prior for Pi.}

\item{domain_alpha}{numeric. Bayes hyperparameter giving prior for domains.}

\item{domain_maxitems}{iinteger. Maximum number of items which can be in a doamin.}

\item{theta_alpha}{numeric. Bayes hyperparemter giving the prior for theta/probabilties.}

\item{domain_proposal_empty}{numeric. Sets how often the domain metropolis proposal function pairs a nonempty domain with an empty domain.}

\item{domain_proposal_swap}{numeric. Sets how often the domain metropolis proposal function swaps items between domains.}

\item{domain_nproposals.}{Sets how many times the domain metropolis propsal function is called each iteration}
}
\description{
Generate hyperparameters
Generate a list of hyperparameters, adding default values when necessary
}
\keyword{internal}
