% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{getStart_thetas}
\alias{getStart_thetas}
\title{Choose starting theta values.}
\usage{
getStart_thetas(mat, classes, hparams)
}
\arguments{
\item{mat}{matrix. Raw data.}

\item{classes}{integer vector. The class of each observation.}

\item{hparams}{list. List of hyperparameters}
}
\description{
Choose starting theta values.
}
\keyword{internal}
