% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{helper_compare_adjmat}
\alias{helper_compare_adjmat}
\title{helper_compare_adjmat}
\description{
Helper function. Returns true if (m1>0) == (m2>0) in all cells.
In other words returns true if both adjacency matrixes have the same connections (ignoring # of possible routes)
}
\keyword{internal}
