% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Hyperparameter::make_pairs}
\alias{Hyperparameter::make_pairs}
\title{Hyperparameter::make_pairs}
\value{
One column per nonzero value. Two rows indicating the row and column position respectively.
}
\description{
Find every nonzero index of domainPriorKnowledgeLog_mat_in. Only searches upper triangle
}
\keyword{internal}
