% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Hyperparameter::set_dataInfo}
\alias{Hyperparameter::set_dataInfo}
\title{Hyperparameter::set_dataInfo}
\description{
Use the raw data we are modeling to set certain hparams settings (e.g. set number of items)
Assumptions: That there are no empty levels especially at end
}
\keyword{internal}
