% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{id2pattern}
\alias{id2pattern}
\title{id2pattern}
\usage{
id2pattern(xpattern, mapvec)
}
\arguments{
\item{xpattern}{Vector of ids each representing a vector}

\item{mapvec}{Vector for converting id to/from vector. 
mapvec is of size equal to the pattern vector length.
mapvec values of 0 produce a PatternVec values of -1 (consider -1 to be NA).}
}
\description{
Convert pattern id to pattern vector
See other instance of id2pattern(.) for details
}
\examples{
Suppose we have a vector of length 4 which we wish to conver to an ID (this function reverses this process).
Suppose the first three indexes take values 0:1 and the last takes values 0:3.
The appropriate mapvec in this situation would be [2,2,2,3] equal to the number of unique values in each position.
If we have a vector of [1,0,1,2] the corresponding id is 1*(1) + 0*(2) + 1*(2*2) + 2*(2*2*2) = 21.
If we have a vector id of 12 we cand find a corresponding vector of [0,0,1,1].
}
\keyword{internal}
