% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{identify_swaps_helper}
\alias{identify_swaps_helper}
\title{For each iteration, identify which class labels are swapped and should be relabeled.
See identify_swaps() for procedure.}
\usage{
identify_swaps_helper(sim_classes, classes_mode, maxitr)
}
\arguments{
\item{sim_classes}{Integer matrix describing what class each observation belongs to in each iteration. There should be one row per observation and one column per iteration as in dlcm$mcmc$classes.}

\item{classes_mode}{Integer vector describing the most likely (mode) class of each observation.}

\item{maxitr}{integer. How many attempts to make.}
}
\description{
For each iteration, identify which class labels are swapped and should be relabeled.
See identify_swaps() for procedure.
}
\keyword{internal}
