% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{insertSorted}
\alias{insertSorted}
\title{insertSorted}
\arguments{
\item{x}{Sorted vector of numbers}

\item{new_value}{new number we wish to insert into x}
}
\description{
Modifies x. Assumes x is sorted. Add new_value to x in correct position based on sort
}
\keyword{internal}
