% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_identifiable}
\alias{is_identifiable}
\title{is_identifiable}
\arguments{
\item{item2superdomainid}{Vector describing which items must be grouped together}

\item{nclass}{Number of classes}

\item{item_nlevels}{Vector with levels of items}
}
\description{
Check if choice of domains is generically identifiable
Uses greedy algorithm. May give false negatives in some cases, but is quick and deterministic
See Allman paper (DOI:10.1214/09-AOS689 Theorem 4.) for criteria used: min(patterns1,nclass)+min(patterns2,nclass)+min(patterns3,nclass) > 2*nclass+2
}
\keyword{internal}
