% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_identifiable_r}
\alias{is_identifiable_r}
\title{is_identifiable_r}
\usage{
is_identifiable_r(item2domainid, nclass, item_nlevels)
}
\arguments{
\item{item2domainid}{Integer matrix with one row per item. For homogeneous DLCM give a single column. In this column items with the same value are interpreted as being in the same domain. For heterogeneous DLCM, provide one column per class. In a single column, items with the same value are assumed to be in the same domain for that class. Putting every item into its own domain is analogous to checking generic identifiability of a traditional LCM.}

\item{nclass}{Number of classes}

\item{item_nlevels}{Vector with levels of each item}
}
\description{
Check if choice of domains is generically identifiable
Uses greedy algorithm. May give false negatives in some cases, but is quick and deterministic
}
