% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{itemid2patterns}
\alias{itemid2patterns}
\title{itemid2patterns}
\usage{
itemid2patterns(pattern_ids, items_ids, item_nlevels)
}
\arguments{
\item{pattern_ids}{ID reprsenting the value of 'filled in' items. (unfilled items get -1)}

\item{items_ids}{ID reprsenting which items (positions) should be 'filled'.}

\item{item_nlevels}{How many possible values each item can take}
}
\description{
Take pattern_id/item_id combinations and form pattern.
}
\examples{
\dontrun{
# Find the pattern of each row in domains from IDs
domain_patterns <- itemid2patterns(
  pattern_ids = dlcm_out$mcmc$domains$pattern_id
  , items_ids = dlcm_out$mcmc$domains$items_id
  , item_nlevels = dlcm_out$hparams$item_nlevels
  )
}
}
