% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDLCM.r
\name{kl_divergence}
\alias{kl_divergence}
\title{Calculates the Kullback-Leibler (KL) divergence for a categorical response.}
\usage{
kl_divergence(true_probs, alternate_probs)
}
\description{
Calculates the Kullback-Leibler (KL) divergence for a categorical response.
}
\keyword{internal}
