% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDLCM.r
\name{kl_max}
\alias{kl_max}
\title{Calculates the maximum KL divergence for a domain under dependence (truth) versus independence (alternate)}
\usage{
kl_max(nlevels)
}
\description{
Calculates the maximum KL divergence for a domain under dependence (truth) versus independence (alternate)
}
\keyword{internal}
