% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{ldomain_prior}
\alias{ldomain_prior}
\title{What is the prior probability of this choice of domains?
Ignores identifiability restrictions}
\usage{
ldomain_prior(x, ndomains, specific_items = FALSE, log = TRUE)
}
\arguments{
\item{x}{IntegerVector. The number of items in each domain. Ok to omit 0's}

\item{ndomains}{Integer. The total number of domains (including empty domains)}

\item{specific_items}{Boolean. 
If FALSE, we look for any domain which produces domains of this size regardless of what specific items they contain.
IF TRUE, we fix which items are in which domain, and calculate the probability of grouping these specific items together.}

\item{log}{Boolean. If TRUE give probability in log scale.}
}
\description{
What is the prior probability of this choice of domains?
Ignores identifiability restrictions
}
