% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{map_get}
\alias{map_get}
\title{map_get}
\arguments{
\item{map}{Map of key:value pairs}

\item{key}{Value you wish to look up}

\item{defaultvalue}{What we should return if key is missing from map}
}
\description{
Look up a value from a map. If no value found return default
}
\keyword{internal}
