% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{match_lookup}
\alias{match_lookup}
\title{match_lookup}
\usage{
match_lookup(lookup, ...)
}
\description{
As base::match(), but with an additional lookup argument.
Where 'x' values match 'table' keys, return the correspond values from 'lookup'.
}
\keyword{export}
