% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM_fit.r
\name{ndomains_singleton_mode}
\alias{ndomains_singleton_mode}
\title{How many domains (ndomains) we we need before 'all singleton domains' are prop-times more likely than 'one 2-item domain with rest singletons'? Using the prior only.
Solves: ldomain_prior(rep(1, nitems), ndomains, specific_items=FALSE, log=TRUE) - ldomain_prior(c(2,rep(1, nitems-2)), ndomains, specific_items=FALSE, log=TRUE) = log(prop)}
\usage{
ndomains_singleton_mode(nitems, prop = 2)
}
\arguments{
\item{nitems}{integer. Number of items in the data.}

\item{prop}{float. prop=1 forces singleton domains to be mode. prop>1 makes singleton domains increasingly frequent. Default is prop=2.}
}
\description{
How many domains (ndomains) we we need before 'all singleton domains' are prop-times more likely than 'one 2-item domain with rest singletons'? Using the prior only.
Solves: ldomain_prior(rep(1, nitems), ndomains, specific_items=FALSE, log=TRUE) - ldomain_prior(c(2,rep(1, nitems-2)), ndomains, specific_items=FALSE, log=TRUE) = log(prop)
}
\keyword{internal}
