% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDLCM.r
\name{predict_class}
\alias{predict_class}
\title{Take new data and predict class membership. Suppose a DLCM has already been fit: 'dlcm<-dependentLCM_fit(...)'.
Note that this will not necessarily match dlcm.summary(dlcm)$classes_cnts and dlcm.summary(dlcm)$classes since this function uses the posterior average response probabilities. This function is intended mainly for new observations not fit in the original model.}
\usage{
predict_class(data_new, item_nlevels, thetas_avg, class_pi_avg)
}
\arguments{
\item{data_new}{matrix. Data matching the existing DLCM run. Column position determines the item. Assumed to already be an integer matrix (e.g. see getStart_matrix() if necessary).}

\item{item_nlevels}{IntegerVector. The number of values each item can take, as in dlcm$hparams$item_nlevels.}

\item{thetas_avg}{Dataframe. Fitted response probabilites as from dlcm.summary(dlcm)$thetas_avg. Only rows with is_mode==1 are used.}

\item{class_pi_avg}{NumericVector. The fitted size of each class as from dlcm.summary(dlcm)$class_pi.}
}
\value{
List with two values. logLiks gives the logged conditional probability of an observation under each class, ignoring class_pi_avg. class_probs gives the probability that each observation belongs with each class, including class_pi_avg.
}
\description{
Take new data and predict class membership. Suppose a DLCM has already been fit: 'dlcm<-dependentLCM_fit(...)'.
Note that this will not necessarily match dlcm.summary(dlcm)$classes_cnts and dlcm.summary(dlcm)$classes since this function uses the posterior average response probabilities. This function is intended mainly for new observations not fit in the original model.
}
