% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print_itr}
\alias{print_itr}
\title{print_itr}
\arguments{
\item{itr}{Number of iterations completed}

\item{nclass}{Number of classes.}

\item{print_itrs}{How frequently to print progress.}
}
\description{
If itr is divisible by print_itrs then print progress to terminal
}
\keyword{internal}
