% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.r
\docType{data}
\name{randomLCA_symptoms_data}
\alias{randomLCA_symptoms_data}
\title{randomLCA_symptoms_data}
\source{
Mihrshahi, S., Peat, J.K., Webb, K., Tovey, R.E., Marks, G.B., Mellis, C.M. and Leeder S.R. (2001) The Childhood Asthma Prevention Study (CAPS): Design and research protocol of a randomized trial for the primary prevention of asthma. Control led Clinical Trials, 22:333-354.
}
\description{
See randomLCA::symptoms. Data was transformed to remove NA values, and to remove the 'Freq' column.
}
\references{
randomLCA::symptoms
}
\keyword{data}
