% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rCategorical}
\alias{rCategorical}
\title{rCategorical}
\arguments{
\item{probs}{Vector of probabilities of each category from 0 to probs.size()-1. Should sum to 1.}
}
\description{
Generate random values from a polytomous categorical distribution
}
\keyword{internal}
