% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rDirichlet}
\alias{rDirichlet}
\title{rDirichlet}
\arguments{
\item{deltas}{vector of dirichlet concentration parameters}
}
\description{
Generate random values from dirichlet distribution
}
\keyword{internal}
