% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{robust_var}
\alias{robust_var}
\title{Robust estimate for variance. Inspired by MCD but fast approximation instead}
\usage{
robust_var(x, alpha)
}
\arguments{
\item{x}{numeric vector. Data we wish to calculate variance of}

\item{alpha}{percentage. Remove the alpha most extreme points before calculating variance.}
}
\description{
Robust estimate for variance. Inspired by MCD but fast approximation instead
}
\keyword{internal}
