% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{sample.df}
\alias{sample.df}
\title{sample.df}
\usage{
sample.df(x, size, ...)
}
\arguments{
\item{x}{The dataframe or matrix we are sampling from}

\item{size}{Number of rows to sample}

\item{...}{Other options per sample.int()}
}
\description{
Samples rows from a dataframe or matrix
}
