% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{sample.integers}
\alias{sample.integers}
\title{sample.integers}
\usage{
sample.integers(x, size, ...)
}
\arguments{
\item{x}{integer vector we are sampling from}

\item{size}{Number of rows to sample}

\item{...}{Other options per sample.int()}
}
\description{
Samples values from an provided integer vector.
Safer version of sample() which has bad behavior if length(x)==1.
}
\keyword{internal}
