% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{set_dimnames}
\alias{set_dimnames}
\title{set_dimnames}
\usage{
set_dimnames(xarray, axis_names)
}
\arguments{
\item{xarray}{The array you wish to name the axis of}

\item{axis_names}{The name of each axis of this array}
}
\description{
Names each axis of your array by naming each row/column axisName#
}
\keyword{internal}
