% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.r
\name{sim_logprob_one}
\alias{sim_logprob_one}
\title{Take the parameters of a simulated lcm (simulate_lcm)
and produce the probability of seeing this observation}
\usage{
sim_logprob_one(xobs, pis, thetas)
}
\arguments{
\item{xobs}{numeric vector. One vector observation}

\item{pis}{numeric vector. How often an observation is in class k.}

\item{thetas}{response probabilities. If a matrix of size Kxlength(pi) then the response probabilites of k bernouli items across length(pi) classes. Otherwise a list of length(pi) (per class) containing lists of length K (per item), containing vectors of response probabilities of each item.}
}
\description{
Take the parameters of a simulated lcm (simulate_lcm)
and produce the probability of seeing this observation
}
\note{
Helper function for sim_logprobs. Currently only supports thetas in list form.
}
\keyword{internal}
