% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{sim_logprobs}
\alias{sim_logprobs}
\title{Take the parameters of a simulated lcm (simulate_lcm)
and produce the probability of seeing each observation given}
\usage{
sim_logprobs(x, pis, thetas)
}
\arguments{
\item{x}{numeric matrix. Data with one row per observation.}

\item{pis}{numeric vector. How often an observation is in class k.}

\item{thetas}{response probabilities. If a matrix of size Kxlength(pi) then the response probabilites of k bernouli items across length(pi) classes. Otherwise a list of length(pi) (per class) containing lists of length K (per item), containing vectors of response probabilities of each item.}
}
\description{
Take the parameters of a simulated lcm (simulate_lcm)
and produce the probability of seeing each observation given
}
\note{
Currently only supports thetas in list form. FUTURE extend to include matrix thetas as well.
}
