% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{simulate_lcm}
\alias{simulate_lcm}
\title{Simulate random latent class model data}
\usage{
simulate_lcm(n, pis, thetas)
}
\arguments{
\item{n}{integer. Number of observations}

\item{pis}{numeric vector. How often an observation is in class k.}

\item{thetas}{response probabilities. If a matrix of size Kxlength(pi) then the response probabilites of k bernouli items across length(pi) classes. Otherwise a list of length(pi) (per class) containing lists of length K (per item), containing vectors of response probabilities of each item.}
}
\description{
Simulate random latent class model data
}
\examples{
set.seed(4)
sim_list1 <- simulate_lcm(
  n = 1000
, pis = c(0.5, 0.5) # 2 classes equally likely
  , thetas = cbind(
    c(rep(0.2, 10), rep(0.8, 10)) # Response probabilities for 20 items in first class
    , c(rep(0.8, 10), rep(0.2, 10))
))

sims_list2 <- simulate_lcm(
 n=1000
, pis=c(0.5, 0.5) # 2 classes equally likely
, thetas_list = list(
  c( rep(list(c(0.8,0.2)), 20)
     , rep(list(c(0.6,0.3,0.05,0.05)), 2)
     , rep(list(c(rep(0.3,3), rep(0.1/5,5))), 1)
  ) # Response probabilities for 23 items in first class
  , c( rep(list(c(0.5,0.5)), 20)
       , rep(list(c(0.3,0.6,0.05,0.05)), 2)
       , rep(list(c(rep(0.4,2), rep(0.2/6,6))), 1)
  )
))
}
