% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theta_alpha_functions}
\alias{theta_alpha_functions}
\alias{theta_alpha_constant}
\alias{theta_alpha_linear}
\alias{theta_alpha_log}
\title{theta_alpha_functions}
\arguments{
\item{theta_alpha}{Scales the strength of the theta dirichlet prior}

\item{ndomainitems}{How many items are in this domain}
}
\description{
Determines how quickly theta decreases as domainitems increase
}
\keyword{internal}
