% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{theta_item_probs}
\alias{theta_item_probs}
\title{Calculate the probabilities of different patterns}
\usage{
theta_item_probs(items, this_sim, itrs)
}
\arguments{
\item{items}{integer vector. Which items should we include? (indexed starting at 1)}

\item{this_sim}{a mcmc simulation. Output from dependentLCM_fit()}

\item{itrs}{integer vector. Which iterations from this_sim should we include?}
}
\value{
Probabilities for each pattern of items
}
\description{
Useful when getting probilities which cross domains.
}
