% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trouble_end}
\alias{trouble_end}
\title{trouble_end}
\arguments{
\item{trouble_id}{ID identifying the executing of the current function taken from trouble_start}

\item{function_name}{name of the current function}
}
\description{
When troubleshooting, run this at end of all functions. Concludes tracking of progress
}
