% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trouble_start}
\alias{trouble_start}
\title{trouble_start}
\arguments{
\item{function_name}{name of the function in question}
}
\value{
An ID identifying this specific execution of this specific function
}
\description{
When troubleshooting, run this at start of all functions. Tracks progress.
}
\keyword{internal}
