% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesContainer::run_init}
\alias{BayesContainer::run_init}
\title{BayesContainer::run_init}
\description{
First initializes, and then does hparams.nitr MCMC steps on all bayes parameters
}
\keyword{internal}
