% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesParameter::domain_resetCounts}
\alias{BayesParameter::domain_resetCounts}
\title{BayesParameter::domain_resetCounts}
\description{
Reset the counts of all domains in given map

Reset the counts in all of BayesParameter's domains
}
\keyword{internal}
