% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesParameter::set_initial}
\alias{BayesParameter::set_initial}
\title{BayesParameter::set_initial}
\description{
Set all BayesParameter properties
See getStart_bayes_params(.) in dependentLCM.r for more details

Set all BayesParameter properties
As other set_initial(.) but supports lists for R compatability
}
\keyword{internal}
