% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDLCM.r
\name{dependence_intensity}
\alias{dependence_intensity}
\title{Takes a single domain. Compare the probabilites under the given probabilities versus under local independence.}
\usage{
dependence_intensity(thetas_avg, response_patterns, items_ids = NULL)
}
\arguments{
\item{thetas_avg}{dataframe. As returned from dlcm.summary()}

\item{response_patterns}{dataframe. As variable inside dlcm.summary(). For each domain/pattern, provides a description.}

\item{items_ids}{integerVector Which domains do we want to process?}
}
\value{
DataFrames with:
\itemize{
\item{"prob_dependent"}{= Probability under local dependence. Same as input 'probs'.}
\item{"prob_marginal"}{= Probability of this response pattern under local independence.}
\item{"odds_ratio"}{= Compares prob_dependent and prob_marginal. Values far from '1' show strong dependence.}
\item{"odds_ratio_str"}{= odds_ratio as a fraction}
}
}
\description{
Takes a single domain. Compare the probabilites under the given probabilities versus under local independence.
}
\keyword{internal}
