% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDLCM.r
\name{dlcm.get_waic_fromagg}
\alias{dlcm.get_waic_fromagg}
\title{Calculate likelihood and WAIC from aggregated information. See dlcm.get_waic() for details.
Warning: Only works if class_loglik_collapsed is saved, as in: dependentLCM_fit(..., save_itrs=c(class_loglik_collapsed=Inf, ...))}
\usage{
dlcm.get_waic_fromagg(dlcm)
}
\description{
Calculate likelihood and WAIC from aggregated information. See dlcm.get_waic() for details.
Warning: Only works if class_loglik_collapsed is saved, as in: dependentLCM_fit(..., save_itrs=c(class_loglik_collapsed=Inf, ...))
}
\keyword{internal}
