% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{dlcm.get_waic_fromraw}
\alias{dlcm.get_waic_fromraw}
\title{Calculate likelihood and WAIC from raw class_loglik. See dlcm.get_waic() for details.}
\usage{
dlcm.get_waic_fromraw(dlcm, itrs = NULL)
}
\description{
Calculate likelihood and WAIC from raw class_loglik. See dlcm.get_waic() for details.
}
\keyword{internal}
