% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{dlcm.summary}
\alias{dlcm.summary}
\title{Get some summary statistics from fitted dependent LCM}
\usage{
dlcm.summary(dlcm, nwarmup = 1000)
}
\arguments{
\item{dlcm}{list. Fitted dependent LCM}

\item{nwarmup}{integer. The first n iterations are considered warmup iterations}
}
\value{
List with:
thetas_avg = average response probabilities for different patterns
domain_items = Which items are commonly grouped together in the same domain
domain_nitems = How often the domains are of size k.
domain_accept = How often our metropolis step accepts its proposal
}
\description{
Get some summary statistics from fitted dependent LCM
}
