% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expSumLog}
\alias{expSumLog}
\title{expSumLog
Same as log(sum(exp(x))), but adjusted to improve precision.
Assumes x are logged values. Calculates sum(e^x) and then converts back to log scale
Handle precision: log(e^a+e^b+e^c) = log(a * (1+e^(b-a)+e^(c-1))) = log(a) + log(1+e^(b-a)+e^(c-1)))}
\usage{
expSumLog(x)
}
\arguments{
\item{x}{numeric vector in log scale}
}
\description{
expSumLog
Same as log(sum(exp(x))), but adjusted to improve precision.
Assumes x are logged values. Calculates sum(e^x) and then converts back to log scale
Handle precision: log(e^a+e^b+e^c) = log(a * (1+e^(b-a)+e^(c-1))) = log(a) + log(1+e^(b-a)+e^(c-1)))
}
\keyword{internal}
