% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDLCM.r
\name{get_class_probs}
\alias{get_class_probs}
\title{For each iteration calculate the probability that a given observation is in each class.
Warning: Only works if class_loglik is saved, as in: dependentLCM_fit(..., save_itrs=c(class_loglik=Inf, ...))}
\usage{
get_class_probs(dlcm)
}
\arguments{
\item{dlcm}{Dependent latent class model outptut from dependentLCM_fit}
}
\description{
For each iteration calculate the probability that a given observation is in each class.
Warning: Only works if class_loglik is saved, as in: dependentLCM_fit(..., save_itrs=c(class_loglik=Inf, ...))
}
