% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{get_jointLikelihood}
\alias{get_jointLikelihood}
\title{For each iteration, calculate the prior probabilities of each parameter and conditional probability of our responses}
\usage{
get_jointLikelihood(dlcm)
}
\arguments{
\item{dlcm}{Dependent latent class model outptut from dependentLCM_fit}
}
\description{
For each iteration, calculate the prior probabilities of each parameter and conditional probability of our responses
}
