% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM_fit.r
\name{getStart_bayes_params}
\alias{getStart_bayes_params}
\title{Generates initial values for bayes parameters}
\usage{
getStart_bayes_params(
  mat,
  hparams,
  class_pi = NULL,
  classes = NULL,
  domains = NULL,
  class_init_method = CLASS_INIT_METHODS[1],
  warmup_dlcm = NULL
)
}
\arguments{
\item{mat}{matrix. The data you wish to analyze. Should be an integer matrix describing nominal data. Each column should take values 0, 1, 2, ..., max-for-this-column. Assumptions: i) In each column the maximum value is achieved for atleast one observation. ii) There are no NAs.}

\item{hparams}{list. List of hyperparameters from getStart_hparams.}

\item{class_pi}{numeric vector size nclass. Initial condition for bayes parameter pi.}

\item{classes}{integer vector size nrow(df). Initial condition for subject classes.}

\item{domains}{list. Initial values for domains/probabilites. Should be of the form:
domains = list(
class1 = list(
  domain1 = list(
    items = c(...) # items in this domain
    , thetas = c(...) # probabilities of each response to these items ordered per dependentLCM::id2pattern()
  )
  , domain2 = list(items=c(...), thetas=c(...))
  , ...
)
, class2 = list(
  domain1 = list(items=c(...), thetas=c(...))
  , domain2 = list(items=c(...), thetas=c(...))
  , ...
)
, ...
)}

\item{class_init_method}{string. Decides how 'classes' is defaulted if NULL. One of "kmodes" or "random" or "random_centers", "random_centers_polar"}

\item{warmup_dlcm}{list. A past DLCM fit (from dependentLCM_fit). The last iteration of the DLCM is used as the Bayes parameters.}
}
\description{
Generates initial values for bayes parameters
}
\keyword{internal}
