% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{getStart_bayes_params}
\alias{getStart_bayes_params}
\title{Generates initial values for bayes parameters}
\usage{
getStart_bayes_params(
  mat,
  hparams,
  class_pi = NULL,
  classes = NULL,
  thetas = NULL,
  class_init_method = "kmodes"
)
}
\arguments{
\item{mat}{matrix. The data you wish to analyze}

\item{hparams}{list. List of hyperparameters from getStart_hparams.}

\item{class_pi}{numeric vector size nclass. Initial condition for bayes parameter pi.}

\item{classes}{integer vector size nrow(df). Initial condition for subject classes.}

\item{thetas}{list. Initial values for thetas/probabilites.}

\item{class_init_method}{string. Decides how 'classes' is defaulted if NULL. One of "kmodes" or "random"}
}
\description{
Generates initial values for bayes parameters
}
\keyword{internal}
