% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{getStart_class_random_centers}
\alias{getStart_class_random_centers}
\title{Choose nclass random centers and put observations in their nearest center}
\usage{
getStart_class_random_centers(mat, hparams)
}
\arguments{
\item{mat}{matrix. Raw data.}

\item{hparams}{list. List of hyperparameters}
}
\description{
Choose nclass random centers and put observations in their nearest center
}
\keyword{internal}
