% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM_fit.r
\name{getStart_hparams}
\alias{getStart_hparams}
\title{Generate hyperparameters
Generate a list of hyperparameters, adding default values when necessary}
\usage{
getStart_hparams(
  nitr,
  df = NULL,
  nitems = NULL,
  nclass = NCLASS,
  ndomains = NULL,
  class2domain = NULL,
  classPi_alpha = CLASSPI_ALPHA,
  domain_maxitems = NULL,
  theta_alpha = THETA_ALPHA,
  domain_proposal_empty = DOMAIN_PROPOSAL_EMPTY,
  domain_nproposals = NULL,
  steps_active = STEPS_ACTIVE,
  save_itrs = SAVE_ITRS,
  domain_theta_prior_type = DOMAIN_THETA_PRIOR_TYPE[1],
  domainPriorKnowledgeLog = NULL,
  print_itrs = Inf
)
}
\arguments{
\item{nitr}{integer. Number of iterations to run the bayes MCMC}

\item{df}{dataframe. The data you wish to analyze. Should describe nominal data. Factor columns work best, but we make an honest effort to interpret any values given (see getStart_matrix()). This will be converted into an integer matrix (see 'matrix' argument). This 'matrix' argument takes precedence over the 'df' argument; only one of these two arguments is needed. Assumes no missing data (NA rows will be dropped).}

\item{nitems}{integer. Number of columns of df.}

\item{nclass}{integer. Number of subject latent classes}

\item{ndomains}{integer. Number of item domains}

\item{class2domain}{integer vector of length nclass. Classes with same value have same domains.
If "HOMO" then defaults to c(0,0,...). If "HET" then defaults to c(0, 1, 2, ...)}

\item{classPi_alpha}{numeric vector. Bayes hyperparameter giving prior for Pi.}

\item{domain_maxitems}{iinteger. Maximum number of items which can be in a domain. Default is 10 (beyond 10 runs slowly).}

\item{theta_alpha}{numeric. Bayes hyperparemter giving the prior for theta/probabilties.}

\item{domain_proposal_empty}{numeric. Sets how often the domain metropolis proposal function pairs a nonempty domain with an empty domain.}

\item{domain_nproposals}{Sets how many times the domain metropolis propsal function is called each iteration}

\item{steps_active}{Named boolean vector of what actions to take during mcmc. If mcmc is skipped then initial values are kept as fixed.
\itemize{
\item{"thetas=TRUE"}{ to do gibbs on response probabilities}
\item{"domains=TRUE"}{ to do metropolis on domains}
\item{"class_pi=TRUE"}{ to do gibbs on class prior}
\item{"classes=TRUE"}{ to do gibbs on class membership}
\item{"identifiable=TRUE"}{ to check generic identifiability conditions of domains}
\item{"likelihood=TRUE"}{ to get the likelihood that each observation is in each class}
\item{"class_collapse=TRUE"}{ to collapse on theta when sampling classes}
}}

\item{save_itrs}{Named numeric vector. Gives the maximum number of MCMC iterations which should be saved. Can set components to infinity to get data from every iteration.
\itemize{
\item{"domains_accept=#}{ for saving domain metropolis accept/reject choices.}
\item{"class_loglik=#}{ saves the probability that a response is observed conditional on each class}
\item{"class_loglik_collapsed=#}{ as class_loglik but after collapsing on response probabilities}
\item{"agg_loglik=#}{ saves aggregate likelihood of each iteration}
\item{"classes=#}{ saves the class vector of each iteration.}
\item{"class_counts=#}{ saves the number of times an observation is in each class.}
\item{"all=#}{ saves everything. If lower, takes precidence over other terms.}
}}

\item{domain_theta_prior_type}{string. Defines what sort of prior is used for domains and theta. One of the following.
\itemize{
\item{"bucket=}{ recommended/default. has moderate regularization on domains. Domain uses "balls in buckets" prior, and theta uses dirichlet prior.}
\item{"patternadjusted=}{ has strong regularization on domains. As bucket, but domain prior is adjusted further to cancel out any theta prior.}
\item{"niave=}{ no regularization on domains. Bad outcomes result. For demonstration purposes only. Assumes all domains are equally likely.}
}}

\item{domainPriorKnowledgeLog}{Numeric. An nitem*nitem upper triangular matrix. Values of zero indicate no prior knowledge of the domain structure. Values greater than zero indicate that this pair of items should be more likely to be in the same domain. Values less than zero indicate that this pair of items should be less likely to be in the same domain.}

\item{print_itrs}{Integer. Optional. Prints progress by printing every print_itrs iterations.}
}
\description{
Generate hyperparameters
Generate a list of hyperparameters, adding default values when necessary
}
\keyword{internal}
