% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentLCM.r
\name{getStart_matrix}
\alias{getStart_matrix}
\title{Convert dataframe of factors to matrix of integers (indexed starting at 0)}
\usage{
getStart_matrix(df)
}
\description{
Convert dataframe of factors to matrix of integers (indexed starting at 0)
}
\keyword{internal}
