% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeDLCM.r
\name{ldomain_prior}
\alias{ldomain_prior}
\title{What is the prior probability of this choice of domains?
Only correct up to a normalizing constant. Ignores identifiability restrictions.}
\usage{
ldomain_prior(
  x,
  ndomains,
  specific_items = FALSE,
  log = TRUE,
  domain_theta_prior_type = "bucket",
  x_npatterns = NULL
)
}
\arguments{
\item{x}{IntegerVector. The number of items in each domain. Ok to omit 0's}

\item{ndomains}{Integer. The total number of domains (including empty domains)}

\item{specific_items}{Boolean. 
If FALSE, we look for any domain which produces domains of this size regardless of what specific items they contain.
IF TRUE, we fix which items are in which domain, and calculate the probability of grouping these specific items together.}

\item{log}{Boolean. If TRUE give probability in log scale.}

\item{domain_theta_prior_type}{String. What domain prior are we using? One of c("bucket", "patternadjusted")}

\item{x_npatterns}{IntegerVector. For domain_theta_prior_type="patternadjusted". For each domain x, how many possible response patterns are there to that domain?}
}
\description{
What is the prior probability of this choice of domains?
Only correct up to a normalizing constant. Ignores identifiability restrictions.
}
